+++ components/cascader/index.spec.js
@@ -74,11 +74,17 @@ describe('Cascader', () => {
         expect(dropdown2.innerHTML).to.matchSnapshot();
 
         await wait(1500);
-        expect(dropdown2.innerHTML).to.matchSnapshot(); 
+        expect(dropdown2.innerHTML).to.matchSnapshot();
+
+        const [item21, item22] = dropdown2.querySelectorAll('.k-item');
+        item22.click();
+        await wait(1);
+        const dropdown3 = getElement('.k-cascader-dropdown');
+        expect(dropdown3.innerHTML).to.matchSnapshot();
     });
 
     it('filter', () => {
-        instance = mount(FilterDemo); 
+        instance = mount(FilterDemo);
 
         const input = instance.element.querySelector('.k-inner');